Made entitys dry
