+++ src/Entity/Address.php
@@ -4,11 +4,13 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
 use PlugAndPay\Sdk\Enum\CountryCode;
+use PlugAndPay\Sdk\Traits\ValidatesFieldMethods;
 
 class Address
 {
+    use ValidatesFieldMethods;
+
     private ?string $city;
     private ?CountryCode $country;
     private ?string $street;
@@ -73,14 +75,5 @@ class Address
         $this->zipcode = $zipcode;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!method_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
