+++ src/Entity/Discount.php
@@ -4,11 +4,13 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
 use PlugAndPay\Sdk\Enum\DiscountType;
+use PlugAndPay\Sdk\Traits\ValidatesFieldMethods;
 
 class Discount
 {
+    use ValidatesFieldMethods;
+
     private float $amount;
     private float $amountWithTax;
     private ?string $code;
@@ -60,14 +62,5 @@ class Discount
         $this->type = $type;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!method_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
