+++ src/Entity/OrderBilling.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
+use PlugAndPay\Sdk\Traits\ValidatesFieldMethods;
 
 class OrderBilling
 {
+    use ValidatesFieldMethods;
+
     private Address $address;
     private Contact $contact;
     private bool $allowEmptyRelations;
@@ -47,14 +49,5 @@ class OrderBilling
         $this->contact = $contact;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!method_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
