+++ src/Entity/PricingTrial.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
+use PlugAndPay\Sdk\Traits\ValidatesFieldMethods;
 
 class PricingTrial
 {
+    use ValidatesFieldMethods;
+
     private float $amount;
     private float $amountWithTax;
     private int $duration;
@@ -46,14 +48,5 @@ class PricingTrial
         $this->duration = $duration;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!method_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
