+++ src/Entity/Stock.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
+use PlugAndPay\Sdk\Traits\HasDynamicFields;
 
 class Stock
 {
+    use HasDynamicFields;
+
     private bool $enabled;
     private bool $hidden;
     private ?int $value;
@@ -46,14 +48,5 @@ class Stock
         $this->value = $value;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!property_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
