+++ src/Entity/SubscriptionBilling.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
+use PlugAndPay\Sdk\Traits\ValidatesFieldMethods;
 
 class SubscriptionBilling
 {
+    use ValidatesFieldMethods;
+
     private bool $allowEmptyRelations;
     private Address $address;
     private Contact $contact;
@@ -81,14 +83,5 @@ class SubscriptionBilling
         $this->paymentOptions = $paymentOptions;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!method_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
