+++ src/Entity/SubscriptionBillingSchedule.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace PlugAndPay\Sdk\Entity;
 
-use BadFunctionCallException;
 use DateTimeImmutable;
 use PlugAndPay\Sdk\Enum\Interval;
+use PlugAndPay\Sdk\Traits\ValidatesFieldMethods;
 
 class SubscriptionBillingSchedule
 {
+    use ValidatesFieldMethods;
+
     protected Interval $interval;
     protected ?int $last;
     protected ?DateTimeImmutable $lastAt;
@@ -98,14 +100,5 @@ class SubscriptionBillingSchedule
         $this->terminationAt = $terminationAt;
 
         return $this;
-    }
-
-    public function isset(string $field): bool
-    {
-        if (!method_exists($this, $field)) {
-            throw new BadFunctionCallException("Field '$field' does not exists");
-        }
-
-        return isset($this->{$field});
     }
 }
